+++ hydromt/workflows/mesh.py
@@ -13,7 +13,6 @@ from xugrid.ugrid import conventions
 
 from .. import gis_utils
 from ..raster import GEO_MAP_COORD
-from .basin_mask import parse_region
 
 logger = logging.getLogger(__name__)
 
@@ -75,7 +74,7 @@ def create_mesh2d(
     mesh2d : xu.UgridDataset
         Generated mesh2d.
     """  # noqa: E501
-    kind, region = parse_region(region, logger=logger)
+
     if kind != "mesh":
         if not isinstance(res, (int, float)):
             raise ValueError("res argument required for kind 'bbox', 'geom'")
